Tried fixing bug in inventory implementation
For those curious, there is a bug in the game right now that registers 3 clicks instead of one, which causes inventory issues.

The player control scheme has the left mouse button mapped to a function called UseItem() in the player controller script. I temporarily disabled the code in the script to try to root out the problem.

Feel free to have a look.